+++ Pandora/Pattern/Functor/Traversable.hs
@@ -1,8 +1,7 @@
 module Pandora.Pattern.Functor.Traversable where
 
-import Pandora.Core.Functor (type (:.), type (:=))
 import Pandora.Pattern.Functor.Covariant (Covariant_)
-import Pandora.Pattern.Functor.Applicative (Applicative, Applicative_)
+import Pandora.Pattern.Functor.Applicative (Applicative_)
 import Pandora.Pattern.Functor.Pointable (Pointable)
 import Pandora.Paradigm.Primary.Algebraic.Product ((:*:))
 
@@ -11,36 +10,13 @@ import Pandora.Paradigm.Primary.Algebraic.Product ((:*:))
 > Let p :: (Pointable t, Pointable g) => t a -> u a
 
 > When providing a new instance, you should ensure it satisfies:
-> * Numeratority of traversing: g . traverse f ≡ traverse (g . f)
-> * Numeratority of sequencing: f . sequence = sequence . comap f
+> * Numeratority of traversing: g . (f -<<---) ≡ (g . f -<<---)
+> * Numeratority of sequencing: f . (identity -<<---)= (identity -<<---) . (f -<$>-)
 > * Preserving point: p (point x) ≡ point x
-> * Preserving apply: f (x <*> y) ≡ f x <*> f y
+> * Preserving apply: f (x -<*>- y) ≡ f x -<*>- f y
 -}
 
-infixl 5 ->>, ->>>, ->>>>, ->>>>>, -<<--, -<<-<<-
-
-class Covariant_ t (->) (->) => Traversable t where
-	{-# MINIMAL (->>) #-}
-	-- | Infix version of 'traverse'
-	(->>) :: (Pointable u (->), Applicative u) => t a -> (a -> u b) -> u :. t := b
-
-	-- | Prefix version of '->>'
-	traverse :: (Pointable u (->), Applicative u) => (a -> u b) -> t a -> u :. t := b
-	traverse f t = t ->> f
-	-- | The dual of 'distribute'
-	sequence :: (Pointable u (->), Applicative u) => t :. u := a -> u :. t := a
-	sequence t = t ->> (\x -> x)
-
-	-- | Infix versions of `traverse` with various nesting levels
-	(->>>) :: (Pointable u (->), Applicative u, Traversable v)
-		=> v :. t := a -> (a -> u b) -> u :. v :. t := b
-	x ->>> f = x ->> (->> f)
-	(->>>>) :: (Pointable u (->), Applicative u, Traversable v, Traversable w)
-		=> w :. v :. t := a -> (a -> u b) -> u :. w :. v :. t := b
-	x ->>>> f = x ->> (->> (->> f))
-	(->>>>>) :: (Pointable u (->), Applicative u, Traversable v, Traversable w, Traversable j)
-		=> j :. w :. v :. t := a -> (a -> u b) -> u :. j :. w :. v :. t := b
-	x ->>>>> f = x ->> (->> (->> (->> f)))
+infixl 5 -<<--, -<<-<<-
 
 class Covariant_ t source target => Traversable_ t source target where
 	(-<<--) :: (Covariant_ u source target, Pointable u target, Applicative_ u (:*:) source target) => source a (u b) -> target (t a) (u (t b))