+++ CHANGELOG.md
@@ -507,3 +507,4 @@
 * Move `Paradigm.Primary.Functor.Product` module to `Paradigm.Primary.Algebraic.Product`
 * Add `Covariant_`, `Pointable` and `Applicative_` constraints to `Traversable_` typeclass
 * Remove `join_` method from `Bindable_` typeclass
+* Remove `Traversable` typeclass